Update nginx-full.rb
When installing the upload module, it meet a error. 
==> ./configure: error: invalid option "--with-upload-module"
I found a error in formula,
at line 104: " ['with-upload',            'with-upload_module',           'Compile with support for Upload module'],"
Exactly, the "Upload module" is a 3rd module, so the "configure" can't accept it.
I had move it up to the last line of the 3rd party modules group from options_data array and set arr[1] to nil.
Please check it.